+++ src/clientfactory/shorthands/__init__.py
@@ -4,7 +4,9 @@ Shorthand reference aliases for extant components
 """
 
 from clientfactory.core import (
-    RequestMethod, Parameter, ParameterType, NestedParameter, Payload
+    RequestMethod, Parameter, ParameterType,
+    NestedParameter, Payload, NestedPayload,
+    ConditionalParameter
 )
 
 from clientfactory.auth import (
@@ -16,7 +18,10 @@ RM = RequestMethod
 P = Parameter
 PT = ParameterType
 NP = NestedParameter
+CP = ConditionalParameter
 PL = Payload
+NPL = NestedPayload
+
 
 # Auth Shorthands
 TS = TokenScheme
@@ -24,12 +29,12 @@ KL = KeyLocation
 
 FULLNAMES = [
     RequestMethod, Parameter, ParameterType, NestedParameter, Payload,
-    TokenScheme, KeyLocation
+    TokenScheme, KeyLocation, NestedPayload, ConditionalParameter
 ]
 
 SHORTHANDS = [
     RM, P, PT, NP, PL,
-    TS, KL
+    TS, KL, NPL, CP
 ]
 
 def genshortdoc(short, full) -> None:
@@ -44,5 +49,5 @@ for short, full in zip(SHORTHANDS, FULLNAMES):
     genshortdoc(short, full)
 
 __all__ = [
-    'RM', 'P', 'PT', 'NP', 'PL', 'TS', 'KL'
+    'RM', 'P', 'PT', 'NP', 'PL', 'TS', 'KL', 'NPL', 'CP'
 ]